worker向dashboard直推热key

+++ dashboard/src/main/java/com/jd/platform/hotkey/dashboard/model/KeyRecord.java
@@ -16,7 +16,7 @@ public class KeyRecord implements Serializable {
     /**
      * 缓存时间
      */
-    private Long duration;
+    private int duration;
 
     /**
      * 来源： SYSTEM 系统探测；USERNAME创建人
@@ -43,7 +43,7 @@ public class KeyRecord implements Serializable {
     public KeyRecord() {
     }
 
-    public KeyRecord(String key,String val, String appName, Long duration,
+    public KeyRecord(String key,String val, String appName, int duration,
                      String source, Integer type,String uuid, Date createTime) {
         this.key = key;
         this.val = val;
@@ -103,11 +103,11 @@ public class KeyRecord implements Serializable {
         this.val = val;
     }
 
-    public Long getDuration() {
+    public int getDuration() {
         return duration;
     }
 
-    public void setDuration(Long duration) {
+    public void setDuration(int duration) {
         this.duration = duration;
     }
